+++ packages/resume/changelog.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Add `replaceStop`, shorthand for `mapStop . const`.
+
 # 0.5.0.0
 
 * Add interceptors.